refactor(#1277): extract OperatorApplicationRepository to keep types barrel under size cap
