+++ packages/api/src/repositories/types.ts
@@ -26,7 +26,6 @@ export type { VehicleDetail } from '@kuruma/shared/types/vehicle-detail'
 export type { Customer, CustomerSort, CustomerWithBookings } from '@kuruma/shared/types/customer'
 
 import type { CoordinateSource } from '@kuruma/shared/db/schema'
-import type { OperatorApplicationStatus } from '@kuruma/shared/enums'
 import type { Customer, CustomerSort, CustomerWithBookings } from '@kuruma/shared/types/customer'
 import type {
   OperatorBookingCounts,
@@ -44,7 +43,6 @@ import type {
   Location,
   MaintenanceLog,
   Operator,
-  OperatorApplication,
   OperatorMembership,
   ProviderInvite,
   Region,
@@ -764,41 +762,6 @@ export type {
   OverviewRepository,
 } from './types-overview'
 
-// #1277: self-serve operator registration data access.
-export interface OperatorApplicationRepository {
-  /** Insert a PENDING application. Throws UNIQUE_VIOLATION on the live-email
-   *  partial index (named OPERATOR_APPLICATION_EMAIL_CONSTRAINT) -> service 409. */
-  create(
-    data: Omit<
-      OperatorApplication,
-      | 'id'
-      | 'status'
-      | 'operatorId'
-      | 'reviewedByUserId'
-      | 'reviewedAt'
-      | 'reviewerNotes'
-      | 'rejectionReason'
-      | 'createdAt'
-      | 'updatedAt'
-    >,
-  ): Promise<OperatorApplication>
-  findById(id: string): Promise<OperatorApplication | undefined>
-  /** Admin queue. Optional status filter; newest first. */
-  list(status?: OperatorApplicationStatus): Promise<OperatorApplication[]>
-  /** Idempotency + claim+link (approval tx step): UPDATE ... SET status='APPROVED',
-   *  operatorId, reviewedByUserId, reviewedAt WHERE id=? AND status='PENDING'
-   *  RETURNING row. Returns undefined when 0 rows matched (already reviewed). */
-  markApprovedIfPending(
-    id: string,
-    operatorId: string,
-    reviewedByUserId: string,
-    reviewedAt: Date,
-  ): Promise<OperatorApplication | undefined>
-  /** REJECTED terminal write. WHERE id=? AND status='PENDING'. */
-  markRejectedIfPending(
-    id: string,
-    reviewedByUserId: string,
-    reviewedAt: Date,
-    rejectionReason: string,
-  ): Promise<OperatorApplication | undefined>
-}
+// #1277: self-serve operator registration data access lives in its own module
+// to keep this barrel under the file-size cap; re-exported for callers.
+export type { OperatorApplicationRepository } from './types-operator-application'
